Encapsulate entry form logic in hook to keep layout component clean
