+++ web/src/components/entry-dialog/DimensionComboBox.tsx
@@ -1,17 +1,16 @@
 import { useQuery } from "@apollo/client";
 import { Autocomplete, FormControl, Grid, TextField } from "@mui/material";
-import { Control, Controller } from "react-hook-form";
+import { Controller, UseFormReturn } from "react-hook-form";
 import { FindDimensionOptionsDocument } from "../../graphql/generated/graphql";
-import { EntryFormSchema } from "./EntryDialog";
+import { EntryFormSchema } from "./useEntryForm";
 
 type DimensionComboBoxProps = {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  control: Control<EntryFormSchema, any>;
+  form: UseFormReturn<EntryFormSchema>;
   name: "product" | "activity" | "issue" | "client";
   title: string;
 };
 
-const DimensionComboBox = ({ control, name, title }: DimensionComboBoxProps) => {
+const DimensionComboBox = ({ form, name, title }: DimensionComboBoxProps) => {
   const { data } = useQuery(FindDimensionOptionsDocument);
 
   const options = data?.findDimensionOptions[name] || [];
@@ -20,7 +19,7 @@ const DimensionComboBox = ({ control, name, title }: DimensionComboBoxProps) => 
     <Grid item xs={12} md={6}>
       <FormControl fullWidth>
         <Controller
-          control={control}
+          control={form.control}
           name={name}
           render={({ field: { value, onChange } }) => {
             return (
