+++ web/src/components/entry-dialog/EntryDialog.tsx
@@ -1,4 +1,3 @@
-import { useMutation } from "@apollo/client";
 import CloseIcon from "@mui/icons-material/Close";
 import {
   Dialog,
@@ -9,122 +8,27 @@ import {
   useMediaQuery,
   useTheme,
 } from "@mui/material";
-import { Dayjs } from "dayjs";
 import { useEffect } from "react";
-import { SubmitHandler, useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
-import dayjs from "../../common/dayjs";
-import {
-  AddWorkdayEntryDocument,
-  Entry,
-  FindWorkdaysDocument,
-  ReplaceWorkdayEntryDocument,
-} from "../../graphql/generated/graphql";
-import { useNotification } from "../global-notification/useNotification";
 import EntryForm from "./EntryForm";
+import useEntryForm, { useEntryProps } from "./useEntryForm";
 
-export type EntryFormSchema = {
-  date: Dayjs;
-  duration: string;
-  description: string;
-  product: string;
-  activity: string;
-  issue: string | null;
-  client: string;
-};
-
-type EntryDialogProps = DialogProps & {
-  editEntry?: Entry;
-  date?: Dayjs;
-  onClose: () => void;
-};
+type EntryDialogProps = DialogProps &
+  useEntryProps & {
+    onClose: () => void;
+  };
 
 const EntryDialog = ({ editEntry, date, onClose, ...props }: EntryDialogProps) => {
   const { t } = useTranslation();
   const theme = useTheme();
   const fullScreen = useMediaQuery(theme.breakpoints.down("md"));
-  const { showSuccessNotification } = useNotification();
+  const { form, onSubmit } = useEntryForm({ editEntry, date });
 
-  const [addWorkdayEntryMutation] = useMutation(AddWorkdayEntryDocument, {
-    refetchQueries: [FindWorkdaysDocument],
-    onCompleted: () => {
-      showSuccessNotification(t("notifications.addEntry.success"));
-    },
-  });
-
-  const [replaceWorkdayEntryMutation] = useMutation(ReplaceWorkdayEntryDocument, {
-    refetchQueries: [FindWorkdaysDocument],
-    notifyOnNetworkStatusChange: true,
-    onCompleted: () => {
-      showSuccessNotification(t("notifications.editEntry.success"));
-    },
-  });
-
-  const defaultValues: EntryFormSchema = {
-    date: date ? dayjs(date) : dayjs(),
-    duration: editEntry?.duration.toString() || "",
-    description: editEntry?.description || "",
-    product: editEntry?.product || "",
-    activity: editEntry?.activity || "",
-    issue: editEntry?.issue || null,
-    client: editEntry?.client || "",
-  };
-
-  const form = useForm<EntryFormSchema>({ defaultValues });
   const {
     handleSubmit,
     reset,
     formState: { isSubmitSuccessful },
   } = form;
-
-  const addWorkday: SubmitHandler<EntryFormSchema> = async (formValues) => {
-    const { date, duration, description, product, activity, issue, client } = formValues;
-
-    await addWorkdayEntryMutation({
-      variables: {
-        entry: {
-          date: date.format("YYYY-MM-DD"),
-          duration: Number(duration),
-          description,
-          product,
-          activity,
-          issue,
-          client,
-        },
-      },
-    });
-  };
-
-  const editWorkday: SubmitHandler<EntryFormSchema> = async (formValues) => {
-    const { date: newDate, duration, description, product, activity, issue, client } = formValues;
-
-    if (!editEntry) {
-      throw new Error("Original entry not given.");
-    }
-
-    await replaceWorkdayEntryMutation({
-      variables: {
-        originalEntry: { key: editEntry.key, date: date?.format("YYYY-MM-DD") },
-        replacementEntry: {
-          date: newDate.format("YYYY-MM-DD"),
-          duration: Number(duration),
-          description,
-          product,
-          activity,
-          issue,
-          client,
-        },
-      },
-    });
-  };
-
-  const onSubmit: SubmitHandler<EntryFormSchema> = async (formValues) => {
-    if (editEntry) {
-      await editWorkday(formValues);
-    } else {
-      await addWorkday(formValues);
-    }
-  };
 
   useEffect(() => {
     if (isSubmitSuccessful) {
