+++ web/src/components/entry-dialog/EntryForm.tsx
@@ -19,8 +19,8 @@ import { Entry } from "../../graphql/generated/graphql";
 import BigDeleteEntryButton from "./BigDeleteEntryButton";
 import DimensionComboBox from "./DimensionComboBox";
 import DurationSlider from "./DurationSlider";
-import { EntryFormSchema } from "./EntryDialog";
 import ResponsiveDatePicker from "./ResponsiveDatePicker";
+import { EntryFormSchema } from "./useEntryForm";
 
 type EntryFormProps = {
   form: UseFormReturn<EntryFormSchema>;
@@ -43,14 +43,10 @@ const EntryForm = ({ reset, onSubmit, editEntry, originalDate, form }: EntryForm
       <Grid container spacing={3}>
         <Grid item xs={12} md={6}>
           <Grid container spacing={2}>
-            <DimensionComboBox control={control} name="product" title={t("entryDialog.product")} />
-            <DimensionComboBox
-              control={control}
-              name="activity"
-              title={t("entryDialog.activity")}
-            />
-            <DimensionComboBox control={control} name="issue" title={t("entryDialog.issue")} />
-            <DimensionComboBox control={control} name="client" title={t("entryDialog.client")} />
+            <DimensionComboBox form={form} name="product" title={t("entryDialog.product")} />
+            <DimensionComboBox form={form} name="activity" title={t("entryDialog.activity")} />
+            <DimensionComboBox form={form} name="issue" title={t("entryDialog.issue")} />
+            <DimensionComboBox form={form} name="client" title={t("entryDialog.client")} />
             <Grid item xs={12}>
               <Controller
                 name="description"
